+++ src/server-dev.ts
@@ -13,4 +13,5 @@ unleash.start({
     },
     enableRequestLogger: true,
     enableOAS: true,
+    dbSession: true,
 });